refactor(ai-service): Step 9 — move agent capability box to app/agents/
The agent framework (planner→executor→synth→evaluator), runners,
sub-agents, tools, and contracts all move out of services/llm/agent/
into a top-level band. Internal structure preserved verbatim — only
the import path changes.

Move:
  app/services/llm/agent/  →  app/agents/

Inside the package: backends/, runners/, sub_agents/, tools/,
contracts.py, evaluators.py, framework.py, __init__.py — all rename
unchanged.

Importers updated (23 files):
  • inside app/agents/  (lots of internal absolute references)
  • app/services/llm/annotation/{flows/,tools/,agents/,followup,formatters}.py
  • app/services/llm/annotation_pipeline.py
  • ui/segment_tabs/{batch.py + components/*_agent_*.py}
  • tests/test_agent_framework.py

Validated:
  • `from app.agents.framework import Agent` etc. resolve cleanly
  • `from app.agents.sub_agents import zoom, label_verifier, describe_graphs`
  • /health 200, app.main imports clean

Pytest still fails to collect — pytest's rootdir auto-detects to
/app/tests/ and app.agents isn't on sys.path. That's a separate
config fix (pytest.ini at the project root), shipped next.
