--- acla_ai_service/app/agents/flows/__init__.py
@@ -1,1 +1,0 @@
-"""Scaffolded for the hexagonal refactor (refactor/hexagonal-v1, Step 1). Empty on purpose; real code lands in later steps."""
