+++ acla_ai_service/app/agents/framework.py
@@ -39,7 +39,7 @@ from typing import Any, Annotated, Callable, Dict, List, Optional, TypedDict
 
 from langgraph.graph import StateGraph, END
 
-from app.services.llm.agent.evaluators import (
+from app.agents.evaluators import (
     AttachmentPool,
     PipelineAttachment,
     merge_pool,
