+++ acla_ai_service/app/agents/runners/__init__.py
@@ -1,1 +1,7 @@
-"""Scaffolded for the hexagonal refactor (refactor/hexagonal-v1, Step 1). Empty on purpose; real code lands in later steps."""
+"""Two execution paradigms, managed separately, sharing one contract.
+
+local.py   LangGraph planner/executor/synthesizer/evaluator subgraph.
+claude.py  Agentic Claude session with MCP tools.
+
+Both consume AgentRequest and emit AgentResponse.
+"""
