--- acla_ai_service/app/services/llm/agent/runners/__init__.py
@@ -1,7 +1,0 @@
-"""Two execution paradigms, managed separately, sharing one contract.
-
-local.py   LangGraph planner/executor/synthesizer/evaluator subgraph.
-claude.py  Agentic Claude session with MCP tools.
-
-Both consume AgentRequest and emit AgentResponse.
-"""
