+++ acla_ai_service/app/services/llm/annotation/tools/__init__.py
@@ -48,7 +48,7 @@ def list_eligible_labels(
             ],
         }
     """
-    from app.services.llm.agent.evaluators import PipelineAttachment
+    from app.agents.evaluators import PipelineAttachment
     from app.skills.label_catalog import find_labels, get_label
 
     def _attach(groups: List[Dict[str, Any]]) -> "PipelineAttachment":
